+++ lib/Payever/ExternalIntegration/Payments/Http/MessageEntity/PaymentDataEntity.php
@@ -16,7 +16,7 @@ namespace Payever\ExternalIntegration\Payments\Http\MessageEntity;
 use Payever\ExternalIntegration\Core\Base\MessageEntity;
 
 /**
- * This class represents Cart Item Entity
+ * This class represents Payment Data Entity
  *
  * @method \DateTime|false  getBirthdate()
  * @method bool             getConditionsAccepted()
@@ -24,10 +24,16 @@ use Payever\ExternalIntegration\Core\Base\MessageEntity;
  * @method string           getFrontendFinishUrl()
  * @method string           getFrontendCancelUrl()
  * @method bool             getForceRedirect()
+ * @method string           getFloor()
+ * @method string           getDoor()
+ * @method string           getBlock()
  * @method self             setConditionsAccepted(bool $conditionsAccepted)
  * @method self             setRiskSessionId(string $riskSessionId)
  * @method self             setFrontendFinishUrl(string $frontendFinishUrl)
  * @method self             setFrontendCancelUrl(string $frontendCancelUrl)
+ * @method self             setFloor(string $floor)
+ * @method self             setDoor(string $door)
+ * @method self             setBlock(string $block)
  */
 class PaymentDataEntity extends MessageEntity
 {
@@ -48,6 +54,21 @@ class PaymentDataEntity extends MessageEntity
 
     /** @var boolean $force_redirect */
     protected $force_redirect;
+
+    /** @var string $organization_name */
+    protected $organization_name;
+
+    /** @var string $street_type */
+    protected $street_type;
+
+    /** @var string $floor */
+    protected $floor;
+
+    /** @var string $door */
+    protected $door;
+
+    /** @var string $block */
+    protected $block;
 
     /**
      * Sets Birthdate
@@ -73,6 +94,54 @@ class PaymentDataEntity extends MessageEntity
     public function setForceRedirect($forceRedirect)
     {
         $this->force_redirect = $forceRedirect;
+
+        return $this;
+    }
+
+    /**
+     * Gets Organization name value
+     *
+     * @return string
+     */
+    public function getOrganizationName()
+    {
+        return $this->organization_name;
+    }
+
+    /**
+     * Sets Organization name value
+     *
+     * @param string $organizationName
+     *
+     * @return $this
+     */
+    public function setOrganizationName($organizationName)
+    {
+        $this->organization_name = $organizationName;
+
+        return $this;
+    }
+
+    /**
+     * Gets Street Type value
+     *
+     * @return string
+     */
+    public function getStreetType()
+    {
+        return $this->street_type;
+    }
+
+    /**
+     * Sets Street Type value
+     *
+     * @param string $streetType
+     *
+     * @return $this
+     */
+    public function setStreetType($streetType)
+    {
+        $this->street_type = $streetType;
 
         return $this;
     }